requirements.txt
torch>=2.0.0
transformers>=4.36.2
Pillow>=9.0.0
numpy>=1.23.0
